ml/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
google-cloud-bigquery==3.25.0
google-cloud-logging==3.11.3
